fix rspec
